### 2023-02-23 #576 @jospueyo
CRediTas: a tiny package to generate CRediT authors statements
Submitting Author Name: Josep Pueyo-Ros
Submitting Author Github Handle: @jospueyo
Repository:  https://github.com/jospueyo/CRediTas
Version submitted: 0.1.0
Submission type: Standard
Version accepted: TBD
---

```
Package: CRediTas
Title: Generate CRediT Author Statements
Version: 0.1.0
    c(
      person(given = "Josep", family = "Pueyo-Ros", email = "josep.pueyo@udg.edu",
      role = c("aut", "cre"), comment = c(ORCID = "0000-0002-1236-5651")))
Description: CRediTas is a tiny package to generate CRediT author statements (https://credit.niso.org/).
    It provides three functions: create a template, read it back and generate the CRediT author statement in a text file.
```

The package facilitates the creation of CRediT author statements, which are mandatory for many journals and encourage good practices in coauthoring scientific publications.

The target audience are authors of scientific publications. There are not scientific applications of this package. It helps in the good practices of open science.

There are not as far as I know

Long lines in test-write_cras.R because I compare a string with a generated text file. If I break the lines I cannot compare both.
Avoid T and F in write_cras.R: I do not use T or F, in line 1, there is "CRediT" which confounds the algorithm.
