+++ vignettes/get_started.Rmd
@@ -1,0 +1,72 @@
+---
+title: "get_started"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{get_started}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
+)
+```
+
+# credit
+
+<!-- badges: start -->
+[![R-CMD-check](https://github.com/jospueyo/credit/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/jospueyo/credit/actions/workflows/R-CMD-check.yaml)
+<!-- badges: end -->
+
+The goal of credit is to facilitate the tedious job of creating [CRediT authors statements](https://credit.niso.org/) for scientific publications.
+
+## Installation
+
+You can install the development version of credit from [GitHub](https://github.com/) with:
+
+``` r
+# install.packages("devtools")
+devtools::install_github("jospueyo/credit")
+```
+
+## Example
+
+The workflow is meant to work with three basic functions. First, we create a template table. It can be created as a `data.frame` and being populated in R. Or as a csv file and being populated in your preferred csv editor.
+
+```{r create_template}
+library(credit)
+
+create_template(authors = c("Alexander Humboldt", "Carl Ritter"), file = tempfile())
+
+cras_table <- create_template(authors = c("Friedrich Ratzel", "Pau Vidal de la Blache", "Élisée Reclus"))
+knitr::kable(cras_table)
+
+```
+As you can see, the table is empty. So you must provide the information of who did what. If you wrote the template to a file, then you can read it back to R as follows:
+
+```{r read_template, eval=FALSE}
+
+cras_table <- read_template(path_to_your_csv_file)
+
+```
+Once the `cras_table` is populated, for instance:
+
+```{r populate_random, echo=FALSE}
+
+cras_table[, 2:ncol(cras_table)] <- sample(0:1, size=3*14, replace = TRUE, prob = c(0.6, 0.4))
+knitr::kable(cras_table)
+
+```
+A text file can be generated following the CRediT author statement format.
+
+```{r}
+
+textfile <- tempfile()
+
+write_cras(cras_table, textfile, markdown = TRUE)
+readLines(textfile)
+
+```
