+++ DESCRIPTION
@@ -15,5 +15,8 @@ URL: https://github.com/jospueyo/credit
 BugReports: https://github.com/jospueyo/credit/issues
 RoxygenNote: 7.2.1
 Suggests: 
+    knitr,
+    rmarkdown,
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
+VignetteBuilder: knitr
