### 2023-02-23 570039c
minor changes in readme
