+++ vignettes/get_started.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "get_started"
+title: "Get started"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{get_started}
@@ -67,6 +67,9 @@ A text file can be generated following the CRediT author statement format.
 textfile <- tempfile()
 
 write_cras(cras_table, textfile, markdown = TRUE)
-readLines(textfile)
 
 ```
+
+If you open the textfile, you will find this:
+
+`r readLines(textfile)`
