### 2023-02-24 a36574b
add sanity checks to write_cras.R and to its test
