+++ DESCRIPTION
@@ -13,7 +13,7 @@ Encoding: UTF-8
 LazyData: true
 URL: https://github.com/jospueyo/credit
 BugReports: https://github.com/jospueyo/credit/issues
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.3
 Suggests: 
     knitr,
     rmarkdown,
