### 2023-02-24 ca2d4ce
fix pkgcheck() issues: sapply and long lines
