+++ vignettes/get_started.Rmd
@@ -39,9 +39,12 @@ The workflow is meant to work with three basic functions. First, we create a tem
 ```{r create_template}
 library(credit)
 
-create_template(authors = c("Alexander Humboldt", "Carl Ritter"), file = tempfile())
+create_template(authors = c("Alexander Humboldt", "Carl Ritter"), 
+                file = tempfile())
 
-cras_table <- create_template(authors = c("Friedrich Ratzel", "Pau Vidal de la Blache", "Élisée Reclus"))
+cras_table <- create_template(authors = c("Friedrich Ratzel", 
+                                          "Pau Vidal de la Blache", 
+                                          "Élisée Reclus"))
 knitr::kable(cras_table)
 
 ```
@@ -56,7 +59,8 @@ Once the `cras_table` is populated, for instance:
 
 ```{r populate_random, echo=FALSE}
 
-cras_table[, 2:ncol(cras_table)] <- sample(0:1, size=3*14, replace = TRUE, prob = c(0.6, 0.4))
+cras_table[, 2:ncol(cras_table)] <- sample(0:1, size=3*14, 
+                                           replace = TRUE, prob = c(0.6, 0.4))
 knitr::kable(cras_table)
 
 ```
