### 2023-02-26 #576 @maurolepore
@jospueyo thanks a lot for your submission!

I'm discussing the scope with the editors board.

To increase the chance they will see the merit of {credit} you might want to unpack in README why the task this package solves is a tedious one. README says it's tedious but it's not immediately obvious how.
