+++ README.md
@@ -10,7 +10,15 @@
 
 The goal of credit is to facilitate the tedious job of creating [CRediT
 authors statements](https://credit.niso.org/) for scientific
-publications.
+publications. Normally, the first author of a scientific paper organizes
+a table in a spreadsheet where all the authors self-state their
+contributions. Often too, it is the first author responsibility to state
+the contributions of all co-authors. However, at the end, the
+information has to be translated to the CRediT statement format of
+“Author 1: roles Authors 2: roles …” which is prone to errors and
+tedious, especially if there are many co-authors. The credit package
+aims to make this easier by providing a template to be filled in form of
+a table (csv) and then converting this table to CRediT statement format.
 
 ## Installation
 
@@ -56,9 +64,9 @@ Once the `cras_table` is populated, for instance:
 
 | Authors                | Conceptualization | Methodology | Software | Validation | Formal analysis | Investigation | Resources | Data curation | Writing - Original Draft | Writing - Review & Editing | Visualization | Supervision | Project administration | Funding acquisition |
 |:-----------------------|------------------:|------------:|---------:|-----------:|----------------:|--------------:|----------:|--------------:|-------------------------:|---------------------------:|--------------:|------------:|-----------------------:|--------------------:|
-| Friedrich Ratzel       |                 0 |           0 |        0 |          0 |               1 |             0 |         0 |             0 |                        0 |                          0 |             1 |           0 |                      0 |                   0 |
-| Pau Vidal de la Blache |                 0 |           1 |        0 |          0 |               1 |             0 |         0 |             1 |                        0 |                          0 |             0 |           0 |                      1 |                   0 |
-| Élisée Reclus          |                 1 |           0 |        0 |          1 |               0 |             1 |         0 |             0 |                        0 |                          0 |             0 |           0 |                      1 |                   1 |
+| Friedrich Ratzel       |                 0 |           1 |        1 |          1 |               1 |             0 |         1 |             1 |                        0 |                          0 |             1 |           0 |                      1 |                   1 |
+| Pau Vidal de la Blache |                 0 |           0 |        1 |          0 |               0 |             0 |         1 |             1 |                        1 |                          1 |             0 |           1 |                      0 |                   0 |
+| Élisée Reclus          |                 0 |           1 |        0 |          0 |               1 |             0 |         0 |             0 |                        0 |                          1 |             1 |           1 |                      1 |                   0 |
 
 A text file can be generated following the CRediT author statement
 format.
@@ -69,9 +77,12 @@ textfile <- tempfile()
 write_cras(cras_table, textfile, markdown = TRUE)
 ```
 
-If you open the textfile, you will find this:
+If you open the text file, you will find this:
 
-**Friedrich Ratzel:** Formal analysis, Visualization **Pau Vidal de la
-Blache:** Methodology, Formal analysis, Data curation, Project
-administration **Élisée Reclus:** Conceptualization, Validation,
-Investigation, Project administration, Funding acquisition
+**Friedrich Ratzel:** Methodology, Software, Validation, Formal
+analysis, Resources, Data curation, Visualization, Project
+administration, Funding acquisition **Pau Vidal de la Blache:**
+Software, Resources, Data curation, Writing - Original Draft, Writing -
+Review & Editing, Supervision **Élisée Reclus:** Methodology, Formal
+analysis, Writing - Review & Editing, Visualization, Supervision,
+Project administration
