### 2023-02-28 #576 @maurolepore
@jospueyo

Thanks for the explanation and for your patience.

I discussed with other editors and find this package to be in scope. I'll start looking for a handling editor.
