### 2023-03-08 #576 @emilyriederer
Hi @jospueyo ! I'm delighted to be handling this from the editorial side. I'm traveling for work W-F so I may be a bit slower to respond, but I thought I'd share some initial feedback to get us started. Thanks again and I look forward to working together.

---

This is a very clean submission, and it feels close to being ready for reviewers! I notice a few small opportunities before we do the hand-off.

- Documentation:
  + The documentation is useful but could be improved to help reviewers navigate
  + The vignette could be enhanced to better show different workflows. For example, the text says "The workflow is meant to work with three basic functions. First, we create a template table. It can be created as a `data.frame` and being populated in R. Or as a csv file and being populated in your preferred csv editor.", but the subsequent code chunk does not show all such options/workflows
  + Please consider making a publicly available pkdown site for your docs (instructions [here](https://devguide.ropensci.org/building.html#website))
- License:
  + The package does not appear to use a standard license at the moment. Please refer to [https://devguide.ropensci.org/building.html#licence](this section) of our docs for more options about licenses

---
