+++ README.Rmd
@@ -25,7 +25,7 @@ The goal of credit is to facilitate the tedious job of creating [CRediT authors 
 
 You can install the development version of credit from [GitHub](https://github.com/) with:
 
-``` {r}
+``` r
 # install.packages("remotes")
 remotes::install_github("jospueyo/credit")
 ```
