### 2023-03-08 #576 @jospueyo
- Documentation:
    - I enhanced the vignette by adding different workflows: write a template to a csv file and define custom roles.
    - I created a public pkgdown site at https://jospueyo.github.io/credit/

- License:
    - I used the functions `use_mit_license()` from {devtools}. Is this correct?
