+++ README.md
@@ -6,6 +6,8 @@
 <!-- badges: start -->
 
 [![R-CMD-check](https://github.com/jospueyo/credit/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/jospueyo/credit/actions/workflows/R-CMD-check.yaml)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/576_status.svg)](https://github.com/ropensci/software-review/issues/576)
 <!-- badges: end -->
 
 The goal of credit is to facilitate the tedious job of creating [CRediT
@@ -64,9 +66,9 @@ Once the `cras_table` is populated, for instance:
 
 | Authors                | Conceptualization | Methodology | Software | Validation | Formal analysis | Investigation | Resources | Data curation | Writing - Original Draft | Writing - Review & Editing | Visualization | Supervision | Project administration | Funding acquisition |
 |:-----------------------|------------------:|------------:|---------:|-----------:|----------------:|--------------:|----------:|--------------:|-------------------------:|---------------------------:|--------------:|------------:|-----------------------:|--------------------:|
-| Friedrich Ratzel       |                 1 |           0 |        0 |          1 |               1 |             0 |         1 |             0 |                        1 |                          1 |             1 |           0 |                      0 |                   0 |
-| Pau Vidal de la Blache |                 0 |           1 |        0 |          0 |               0 |             0 |         0 |             0 |                        0 |                          0 |             1 |           1 |                      0 |                   0 |
-| Élisée Reclus          |                 1 |           1 |        1 |          0 |               1 |             1 |         0 |             0 |                        1 |                          0 |             1 |           0 |                      0 |                   1 |
+| Friedrich Ratzel       |                 1 |           1 |        0 |          0 |               0 |             1 |         1 |             0 |                        0 |                          0 |             1 |           0 |                      0 |                   0 |
+| Pau Vidal de la Blache |                 1 |           1 |        0 |          0 |               0 |             0 |         1 |             0 |                        0 |                          1 |             0 |           0 |                      0 |                   0 |
+| Élisée Reclus          |                 1 |           0 |        1 |          1 |               0 |             1 |         1 |             0 |                        1 |                          0 |             0 |           1 |                      1 |                   0 |
 
 A text file can be generated following the CRediT author statement
 format.
@@ -79,9 +81,8 @@ write_cras(cras_table, textfile, markdown = TRUE)
 
 If you open the text file, you will find this:
 
-**Friedrich Ratzel:** Conceptualization, Validation, Formal analysis,
-Resources, Writing - Original Draft, Writing - Review & Editing,
-Visualization **Pau Vidal de la Blache:** Methodology, Visualization,
-Supervision **Élisée Reclus:** Conceptualization, Methodology, Software,
-Formal analysis, Investigation, Writing - Original Draft, Visualization,
-Funding acquisition
+**Friedrich Ratzel:** Conceptualization, Methodology, Investigation,
+Resources, Visualization **Pau Vidal de la Blache:** Conceptualization,
+Methodology, Resources, Writing - Review & Editing **Élisée Reclus:**
+Conceptualization, Software, Validation, Investigation, Resources,
+Writing - Original Draft, Supervision, Project administration
