### 2023-03-17 #576 @jospueyo
> Hi @jospueyo - I'm delighted to introduce @msperlin, as seasoned rOpenSci package dev, as your seond reviewer

Hi! I'm excited about this process! A lot to learn!
