### 2023-03-20 #576 @msperlin
---

`credit` is a tiny package that automates the process of writing a contribution matrix in the [CRedit](https://credit.niso.org/) format, and can be very useful for academic work, specially in large research groups.

Overall, I liked the package and it worked fine in my machine. But, in order to improve it, I do have some suggestions:

#### package name

The name `credit` might confuse some users. At face value, an R package named _credit_ is most likely to be related to  mathematical finance and credit markets, and not to the CRedit reference system. My suggestion is to think about  another, more intuitive, name. A couple suggestions: CReditR, contribR,

#### possible integration with other packages

There are many R packages for [reproducible research](https://cran.r-project.org/web/views/ReproducibleResearch.html), including handling research metadata such as bibtext references.

I wonder if the authors have checked for any possible integration with these packages. I don't have a direct suggestion, but might be worth the time to check for it.

#### use template for roles

The roles used in the code, as in "Conceptualization", "Methodology" and so on, should be defined in a single place. Do notice that you hard coded (wrote in a script) the roles in function `create_template` and in test `test-read_template`.

There are many ways to do it, but the simples is just using a function:

```
get_roles <- function() {

  roles <- c(
  "Conceptualization",
  "Methodology",
  "Software",
  "Validation",
  "Formal analysis",
  "Investigation",
  "Resources",
  "Data curation",
  "Writing - Original Draft",
  "Writing - Review & Editing",
  "Visualization",
  "Supervision",
  "Project administration",
  "Funding acquisition"
)

  return(roles)
}
```

With some simple webscrapping, you could use `testthat` to  check if the template in the package is identical to the template from the [site](https://credit.niso.org/). Whenever the site adds or removes a role, your CHECK process will show it.

#### Function and argument naming

From the ropensci packaging [guideline](https://devguide.ropensci.org/building.html#function-and-argument-naming), I strongly advise to use the object_verb() format (and not the verb_object() pattern) for function names.

#### use cli messages

Package `cli` is a great way to communicate to the user the output of all functions. It can also include clickable links to local files. This can be very helpful in function `write_cras()`, where the users can simply click in the prompt to open the resulting text file.

#### Functions

##### create_template.R

In the description, you should add a link to the [CRedit](https://credit.niso.org/) and also improve the text. This is your main function, and most likely where users will spend more time. You should try to document it as much as possible.

##### read_template.R

Since this function is reading a template, you should add some test to check if the information in the file has the expected structure (column names, number of columns, ..). If not, issuing a _warning_ might be a good policy, letting the user know that the expected format was not found.

##### write_cras.R

In line 34, I suggest you add an input option for the user to drop (or not) authors without contributions. I suspect that, in large research groups, might be best to also make it explicit that some people did not contribute.

The description of the function was not found. Since this is a tiny package, I suggest to add a nice and comprehensive description of the function.
