+++ README.md
@@ -1,35 +1,35 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# credit
+# CRediTas
 
 <!-- badges: start -->
 
-[![R-CMD-check](https://github.com/jospueyo/credit/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/jospueyo/credit/actions/workflows/R-CMD-check.yaml)
+[![R-CMD-check](https://github.com/jospueyo/CRediTas/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/jospueyo/CRediTas/actions/workflows/R-CMD-check.yaml)
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/576_status.svg)](https://github.com/ropensci/software-review/issues/576)
 <!-- badges: end -->
 
-The goal of credit is to facilitate the tedious job of creating [CRediT
-authors statements](https://credit.niso.org/) for scientific
+The goal of CRediTas is to facilitate the tedious job of creating
+[CRediT authors statements](https://credit.niso.org/) for scientific
 publications. Normally, the first author of a scientific paper organizes
 a table in a spreadsheet where all the authors self-state their
 contributions. Often too, it is the first author responsibility to state
 the contributions of all co-authors. However, at the end, the
 information has to be translated to the CRediT statement format of
 “Author 1: roles Authors 2: roles …” which is prone to errors and
-tedious, especially if there are many co-authors. The credit package
+tedious, especially if there are many co-authors. The CRediTas package
 aims to make this easier by providing a template to be filled in form of
 a table (csv) and then converting this table to CRediT statement format.
 
 ## Installation
 
-You can install the development version of credit from
+You can install the development version of CRediTas from
 [GitHub](https://github.com/) with:
 
 ``` r
 # install.packages("remotes")
-remotes::install_github("jospueyo/credit")
+remotes::install_github("jospueyo/CRediTas")
 ```
 
 ## Example
@@ -40,7 +40,7 @@ populated in R. Or as a csv file and being populated in your preferred
 csv editor.
 
 ``` r
-library(credit)
+library(CRediTas)
 
 create_template(authors = c("Alexander Humboldt", "Carl Ritter"), file = tempfile())
 
@@ -66,9 +66,9 @@ Once the `cras_table` is populated, for instance:
 
 | Authors                | Conceptualization | Methodology | Software | Validation | Formal analysis | Investigation | Resources | Data curation | Writing - Original Draft | Writing - Review & Editing | Visualization | Supervision | Project administration | Funding acquisition |
 |:-----------------------|------------------:|------------:|---------:|-----------:|----------------:|--------------:|----------:|--------------:|-------------------------:|---------------------------:|--------------:|------------:|-----------------------:|--------------------:|
-| Friedrich Ratzel       |                 1 |           1 |        0 |          0 |               0 |             1 |         1 |             0 |                        0 |                          0 |             1 |           0 |                      0 |                   0 |
-| Pau Vidal de la Blache |                 1 |           1 |        0 |          0 |               0 |             0 |         1 |             0 |                        0 |                          1 |             0 |           0 |                      0 |                   0 |
-| Élisée Reclus          |                 1 |           0 |        1 |          1 |               0 |             1 |         1 |             0 |                        1 |                          0 |             0 |           1 |                      1 |                   0 |
+| Friedrich Ratzel       |                 0 |           1 |        1 |          0 |               0 |             0 |         0 |             1 |                        0 |                          0 |             0 |           0 |                      0 |                   1 |
[…]
