### 2023-03-21 #576 @jospueyo
BTW, following the first comment, I changed the package name to CRediTas, and consequently, the repository name. How is this affecting this issue in terms of links or other problems?
