### 2023-03-21 #576 @msperlin
> Hi! I opened separate issues in the package repository. Is it fine if I report the changes there? Or should I do here?

I'm fine with the changes as issues.

> BTW, following the first comment, I changed the package name to CRediTas, and consequently, the repository name. How is this affecting this issue in terms of links or other problems?

I'm not sure of the consequences. Perhaps you should edit the first message. Or, resubmit the package.
@emilyriederer can you help?
