### 2023-03-22 25d6844
close issue #2 possible integration with other packages
