+++ R/write_cras.R
@@ -10,7 +10,7 @@
 #' If FALSE, a warning is triggered in case any authors is dropped out.
 #' @return A text file with the CRediT authors statement or, if file is NULL
 #' (default), a string with the statement that can be used in a Rmarkdown or
-#' quarto document using inline code `r write_cras(cras_table, markdown = TRUE)`
+#' quarto document using inline code: \code{`r write_cras(cras_table, markdown = TRUE)`}
 #' @examples
 #' # Generate a template and populate it (randomwly for this example)
 #' cras_table <- create_template(authors = c("Josep Maria", "Jane Doe"))
@@ -68,5 +68,5 @@ write_cras <- function(cras_table,
   if(file.exists(file) && isFALSE(overwrite)) stop("The file already exists")
 
   writeLines(cras, file)
-  invisible()
+  invisible(cras)
 }
