+++ man/write_cras.Rd
@@ -30,7 +30,7 @@ If FALSE, a warning is triggered in case any authors is dropped out.}
 \value{
 A text file with the CRediT authors statement or, if file is NULL
 (default), a string with the statement that can be used in a Rmarkdown or
-quarto document using inline code `r write_cras(cras_table, markdown = TRUE)`
+quarto document using inline code: \code{`r write_cras(cras_table, markdown = TRUE)`}
 }
 \description{
 Write CRediT author statement
