+++ R/roles_get.R
@@ -1,0 +1,21 @@
+#' Get default roles for CRediT
+#'
+
+roles_get <- function() {
+    c(
+    "Conceptualization",
+    "Methodology",
+    "Software",
+    "Validation",
+    "Formal Analysis",
+    "Investigation",
+    "Resources",
+    "Data curation",
+    "Writing - original draft",
+    "Writing - review & editing",
+    "Visualization",
+    "Supervision",
+    "Project administration",
+    "Funding acquisition"
+  )
+}
