+++ tests/testthat/test-read_template.R
@@ -1,22 +1,7 @@
 file <- tempfile()
 create_template(authors = c("A", "B"), file = file)
 
-roles <- c(
-  "Conceptualization",
-  "Methodology",
-  "Software",
-  "Validation",
-  "Formal analysis",
-  "Investigation",
-  "Resources",
-  "Data curation",
-  "Writing - Original Draft",
-  "Writing - Review & Editing",
-  "Visualization",
-  "Supervision",
-  "Project administration",
-  "Funding acquisition"
-)
+roles <- roles_get()
 
 test_that("read_template works", {
   expect_s3_class(read_template(file), "data.frame")
