+++ DESCRIPTION
@@ -17,6 +17,7 @@ RoxygenNote: 7.2.3
 Suggests: 
     knitr,
     rmarkdown,
+    rvest,
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
 VignetteBuilder: knitr
