### 2023-03-22 a866cc9
close #4 functions renamed to object_verb()
