+++ NEWS.md
@@ -15,8 +15,13 @@ quarto. When file is NULL (new default), instead of writing a file, the
 function returns a string with the CRediT Author Statement. It works well with 
 inline code in rmarkdown and quarto documents: 
 `r cras_write(cras_table, markdown = TRUE)` (#2)
+- `cras_write` includes a new argument `drop_authors`, if set to `FALSE`, the authors without contributions are not removed from the statement.
 
 ## Minor improvements and fixes
+
 - The roles are created using the function roles_get() instead of hard coded.
 - The former allows to check if authors’ roles are up-to-date comparing with
 [CRediT] (https://credit.niso.org/)
+- Improvements on the documentation of all functions
+- In `cras_write`, the default of `markdown` argument is `TRUE`.
+- Added some checks to `template_read` to ensure that the format is correct.

[…]
