### 2023-03-23 #576 @jospueyo
I addressed all comments of @msperlin. You can see all changes in news: https://github.com/jospueyo/CRediTas/blob/master/NEWS.md
