### 2023-03-23 #576 @msperlin
> I addressed all comments of @msperlin. You can see all changes in news: https://github.com/jospueyo/CRediTas/blob/master/NEWS.md

Thanks @jospueyo. Great work. LEt me know if I can help any further.
