### 2023-04-04 jospueyo/CRediTas#7 @jospueyo
fix CSS of pkgdown
docs/ (important) fix the link to the CSS of the package dedicated documentation page (https://jospueyo.github.io/CRediTas/) appears to be broken as the pkdown package layout is missing
