### 2023-04-04 jospueyo/CRediTas#8 @jospueyo
Add @description to template_create()
template_create() (important): the @description roxygen2 package tag is missing. This originates a very long headline in ?template_create
