### 2023-04-04 jospueyo/CRediTas#9 @jospueyo
Add fix(cras_table) to readme and get_started
README (not important): in the Example section, add fix(cras_table) (right after knitr::kable(cras_table)) to edit the authorship contribution dataframe in a spreadsheet-like manner directly from within R or RStudio
