### 2023-04-04 jospueyo/CRediTas#10 @jospueyo
Add realted packages tp README
README (not important): add a Related packages section with links to packages covering related use cases (eg. [marton-balazs-kovacs/tenzing](https://github.com/marton-balazs-kovacs/tenzing/) and [openbiox/contribution](https://github.com/openbiox/contribution)) at the end of the file
