### 2023-04-04 jospueyo/CRediTas#11 @jospueyo
Use missing instead of is.null in cras_write()
allow the file parameter to be empty in cras_write() (not important): check if the file parameter has been assigned with if (missing(file)) instead of if(is.null(file))
