### 2023-04-04 jospueyo/CRediTas#12 @jospueyo
Reduce the scope of template_create()
reduce the scope of template_create() and remove template_read() (not important)

writing/reading output files with write.csv2()/read.csv2() (and/or the readxl package) could be shown in the README and in a vignette. template_create() could be even shorter:
