+++ DESCRIPTION
@@ -5,7 +5,10 @@ Version: 0.1.0
 Authors@R: 
     c(
       person(given = "Josep", family = "Pueyo-Ros", email = "josep.pueyo@udg.edu", 
-      role = c("aut", "cre"), comment = c(ORCID = "0000-0002-1236-5651")))
+      role = c("aut", "cre"), comment = c(ORCID = "0000-0002-1236-5651")),
+      person(given = "Marcelo S.", family = "Perlin", role = "rev", comment = c(github = "@msperlin")),
+      person(given = "João", family = "Martins", role = "rev", comment = c(github = "@zambujo"))
+      )
 Description: CRediTas is a tiny package to generate CRediT author statements (https://credit.niso.org/). 
     It provides three functions: create a template, read it back and generate the CRediT author statement in a text file.
 License: MIT + file LICENSE
