+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: ~
+url: https://jospueyo.github.io/CRediTas/
 template:
   bootstrap: 5
 
