### 2023-04-04 c474702
close #9 add fix() to readme and get_started
