### 2023-04-04 jospueyo/CRediTas#9 @jospueyo
@zambujo. I added the exmaple with `fix` to readme and get_started vignette. I set up as eval = FALSE because I think this only works in interactive environments. Is it right?

BTW, I loved this, I didn't know this function. It has some issues with reproducibility but it is very handy for some specific tasks.
