+++ README.Rmd
@@ -85,3 +85,7 @@ If you open the text file, you will find this:
 `r cras_write(cras_table, markdown = TRUE, quiet = TRUE)`
 
 
+## Related packages
+
+- [tenzing](https://github.com/marton-balazs-kovacs/tenzing/): Tenzing, an easy-to-use web-based app, allows researchers to generate reports about the contribution of each team member on a project using CRediT, for insertion into their manuscripts and for publishers to potentially incorporate into article metadata.
+- [contribution](https://github.com/openbiox/contribution): The goal of contribution is to generate contribution table for credit assignment in a project. This is inspired by Nick Steinmetz (see twitter https://twitter.com/SteinmetzNeuro/status/1147241138291527681).
