### 2023-04-04 #576 @jospueyo
## Responses to @zambujo:

- I added both reviewers (do you agree @msperlin?) https://github.com/jospueyo/CRediTas/issues/13
-   I fixed the CSS links in pkgdown https://github.com/jospueyo/CRediTas/issues/7
- I added the description tag to template_create https://github.com/jospueyo/CRediTas/issues/8
- I added the `fix` function in readme and get_started vignette https://github.com/jospueyo/CRediTas/issues/9
- I added related packages to README as suggested https://github.com/jospueyo/CRediTas/issues/10
- I changed `is.null(file)` by `missing(file)` in `cras_write`. https://github.com/jospueyo/CRediTas/issues/11
- I modified `template_create` following your suggestions. I justified why I prefer keeping the i/o functions as part of the package https://github.com/jospueyo/CRediTas/issues/12
