### 2023-04-04 #576 @jospueyo
For the long run suggestions, I will keep those in mind. I liked both ideas from @zambujo. The second one is especially interesting for epistemological research. I also have in mind an option to facilitate copy&paste in MS word, currently the **bold** is missed on the way to MS word.

Any contributions will be welcome! ;)
