### 2023-04-04 #576 @msperlin
> ## Responses to @zambujo:
> * I added both reviewers (do you agree @msperlin?) [Add reviewers to DESCRIPTION jospueyo/CRediTas#13](https://github.com/jospueyo/CRediTas/issues/13)
> * I fixed the CSS links in pkgdown [fix CSS of pkgdown jospueyo/CRediTas#7](https://github.com/jospueyo/CRediTas/issues/7)
> * I added the description tag to template_create [Add @description to template_create() jospueyo/CRediTas#8](https://github.com/jospueyo/CRediTas/issues/8)
> * I added the `fix` function in readme and get_started vignette [Add fix(cras_table) to readme and get_started jospueyo/CRediTas#9](https://github.com/jospueyo/CRediTas/issues/9)
> * I added related packages to README as suggested [Add realted packages tp README jospueyo/CRediTas#10](https://github.com/jospueyo/CRediTas/issues/10)
> * I changed `is.null(file)` by `missing(file)` in `cras_write`. [Use missing instead of is.null in cras_write() jospueyo/CRediTas#11](https://github.com/jospueyo/CRediTas/issues/11)
> * I modified `template_create` following your suggestions. I justified why I prefer keeping the i/o functions as part of the package [Reduce the scope of template_create() jospueyo/CRediTas#12](https://github.com/jospueyo/CRediTas/issues/12)

I agree.
