### 2023-04-05 #576 @zambujo
Many thanks @jospueyo: great to see the changes.

If i/o functionality is to be kept, then I'd like to extend my suggestion of not setting `file = NULL` by default to `template_create()` and `template_read()` (replacing `is.null(file)` with `missing(file)`).

In `template_create()`, I'd also suggest passing file as the second parameter.

And to help beginners, it might be a good idea to include path nomalization with `normalizePath(file)`.

Eg.:

```r
template_create <- function(authors, file, roles = roles_get()) {
  df <- data.frame(Authors = authors)
  mat <- matrix(0, nrow = length(authors), ncol = length(roles))
  colnames(mat) <- roles
  df <- cbind(df, mat)
  if (is.missing(file)) return(df)

  file <- normalizePath(file)
  write.csv2(df, file, row.names = FALSE)
  invisible()
}
```

Last, consider updating the error message [in `cras_write.R`, line 82](https://github.com/jospueyo/CRediTas/blob/4dd185a42e3e0e93cd972ebc01e4b3778bed1b4e/R/cras_write.R#L82). 😉
