### 2023-04-08 #576 @emilyriederer
@msperlin - It looks like you're satisfied with the responses to your feedback? If so, can you please document your approval with our {Final Approval Form](https://devguide.ropensci.org/approval2template.html) ?

(Same for @zambujo once you're satisfied with the discussion!)
