+++ NEWS.md
@@ -7,6 +7,7 @@
     - `create_template` is `template_create`
     - `read_template` is `template_read`
     - `write_cras` is `cras_write`
+- `file` argument in `cras_write` and in `template_create` no longer default to `NULL`, they are not provided by default. If missing, they behave like the old `NULL` value. 
 
 ## New features
 
@@ -14,7 +15,7 @@
 quarto. When file is NULL (new default), instead of writing a file, the 
 function returns a string with the CRediT Author Statement. It works well with 
 inline code in rmarkdown and quarto documents: 
-`r cras_write(cras_table, markdown = TRUE)` (#2)
+`r cras_write(cras_table, markdown = TRUE)` (#2
 - `cras_write` includes a new argument `drop_authors`, if set to `FALSE`, the authors without contributions are not removed from the statement. (#6)
 
 ## Minor improvements and fixes
@@ -25,3 +26,9 @@ inline code in rmarkdown and quarto documents:
 - Improvements on the documentation of all functions (#6)
 - In `cras_write`, the default of `markdown` argument is `TRUE`. 
 - Added some checks to `template_read` to ensure that the format is correct. (#6)
+- Added reviewers of ROpenScience to `DESCRIPTION`. 
+- CSS fixed in pkgdown.
+- Description added to `template_create` documentation.
+- Option to use `fix()` to edit the template is added to `README` and `get_started` vignette.
+- Related packages added to `README`.
+- `template_create` is more performant thanks to using `matrix` instead of `data.frame` to create the empty contributions.
