### 2023-04-11 #576 @yabellini
Hi, @jospueyo (rOpenSci Community Manager here). If you want to, I can invite you to our Slack workspace.  I will need your email.  You can send it to yabellini@ropensci.org.
